fix(prompt-gate): share message reservations
